requirements.txt
accelerate
peft
bitsandbytes
transformers
trl
datasets

faiss-cpu
langchain
python-dotenv
tiktoken
sentence_transformers
einops